# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Risk Control API",
        "description": "API para sistema de control de riesgo en operaciones de trading",
        "contact": {
            "email": "soporte@riskcontrol.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/accounts": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Listar cuentas",
                "description": "Obtener lista de cuentas",
                "operationId": "96daed062b209799ebd8c7e84dc97aed",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado de cuenta",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "enable",
                                "disable"
                            ]
                        }
                    },
                    {
                        "name": "trading_status",
                        "in": "query",
                        "description": "Filtrar por estado de trading",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "enable",
                                "disable"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de cuentas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/PaginatedResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Account"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Crear cuenta",
                "description": "Crear nueva cuenta",
                "operationId": "b3234b57f04a5c23b5d70ac83678df8c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "trading_status",
                                    "status"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "integer",
                                        "example": 21002025
                                    },
                                    "trading_status": {
                                        "type": "string",
                                        "enum": [
                                            "enable",
                                            "disable"
                                        ],
                                        "example": "enable"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "enable",
                                            "disable"
                                        ],
                                        "example": "enable"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cuenta creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Account created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Account"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/accounts/{id}": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Obtener cuenta por ID",
                "description": "Obtener cuenta por ID",
                "operationId": "263a1e8964630a13950ec2af14d9d9c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la cuenta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cuenta obtenida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Account"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Actualizar cuenta",
                "description": "Actualizar cuenta",
                "operationId": "f0a94d0d1e9d6f4d58fcd33b60baf45a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la cuenta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "login": {
                                        "type": "integer",
                                        "example": 21002025
                                    },
                                    "trading_status": {
                                        "type": "string",
                                        "enum": [
                                            "enable",
                                            "disable"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "enable",
                                            "disable"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cuenta actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Account updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Account"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Eliminar cuenta",
                "description": "Eliminar cuenta",
                "operationId": "ef1e151477fc1ebb2938b30e27606342",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la cuenta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cuenta eliminada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Account deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
        "/accounts/{id}/disable-trading": {
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Deshabilitar trading",
                "description": "Deshabilitar trading de cuenta",
                "operationId": "2dde123440d7264e61add46b9d7c48e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la cuenta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trading deshabilitado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Trading disabled for account"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Account"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "PaginatedResponse": {
                "title": "Respuesta Paginada",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "links": {
                        "type": "object"
                    },
                    "meta": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "Account": {
                "title": "Cuenta de Trading",
                "description": "Representa una cuenta de trading en el sistema",
                "required": [
                    "login",
                    "trading_status",
                    "status"
                ],
                "properties": {
                    "id": {
                        "description": "Identificador único",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "login": {
                        "description": "Identificador numérico de login para la cuenta de trading",
                        "type": "integer",
                        "example": 21002025
                    },
                    "trading_status": {
                        "description": "Estado de actividad de trading",
                        "type": "string",
                        "enum": [
                            "enable",
                            "disable"
                        ],
                        "example": "enable"
                    },
                    "status": {
                        "description": "Estado general de la cuenta",
                        "type": "string",
                        "enum": [
                            "enable",
                            "disable"
                        ],
                        "example": "enable"
                    },
                    "created_at": {
                        "description": "Fecha y hora de creación de la cuenta",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha y hora de última actualización",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "description": "Fecha y hora de eliminación suave (si aplica)",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}